refactor: global 패키지 리패키징 및 리팩토링
