+++ src/main/java/io/wisoft/wasabi/domain/board/BoardController.java
@@ -1,8 +1,8 @@
 package io.wisoft.wasabi.domain.board;
 
 import io.wisoft.wasabi.domain.board.dto.*;
-import io.wisoft.wasabi.global.config.common.annotation.Anyone;
-import io.wisoft.wasabi.global.config.common.annotation.MemberId;
+import io.wisoft.wasabi.global.config.web.resolver.Anyone;
+import io.wisoft.wasabi.global.config.web.resolver.MemberId;
 import io.wisoft.wasabi.global.config.web.response.Response;
 import io.wisoft.wasabi.global.config.web.response.ResponseType;
 import jakarta.validation.Valid;
