+++ src/main/java/io/wisoft/wasabi/domain/board/dto/UploadImageRequest.java
@@ -1,6 +1,6 @@
 package io.wisoft.wasabi.domain.board.dto;
 
-import io.wisoft.wasabi.global.config.common.annotation.FileExtension;
+import io.wisoft.wasabi.global.config.web.validator.FileExtension;
 import org.springframework.web.multipart.MultipartFile;
 
 public record UploadImageRequest(
