+++ src/main/java/io/wisoft/wasabi/domain/like/LikeController.java
@@ -4,7 +4,7 @@ import io.wisoft.wasabi.domain.like.dto.CancelLikeResponse;
 import io.wisoft.wasabi.domain.like.dto.GetLikeResponse;
 import io.wisoft.wasabi.domain.like.dto.RegisterLikeRequest;
 import io.wisoft.wasabi.domain.like.dto.RegisterLikeResponse;
-import io.wisoft.wasabi.global.config.common.annotation.Anyone;
+import io.wisoft.wasabi.global.config.web.resolver.Anyone;
 import io.wisoft.wasabi.global.config.web.response.Response;
 import io.wisoft.wasabi.global.config.web.response.ResponseType;
 import jakarta.validation.Valid;
