+++ src/main/java/io/wisoft/wasabi/global/config/common/Const.java
@@ -36,4 +36,20 @@ public final class Const {
     public static final String IMAGE_EXTENSION_PNG = "image/png";
     public static final String IMAGE_EXTENSION_JPG = "image/jpg";
     public static final String CONTENT_TYPE_IMAGE = "image/";
+
+    public static String[] getClientIpHeaders() {
+        return new String[] {
+            Const.X_FORWARDED_FOR,
+            Const.PROXY_CLIENT_IP,
+            Const.WL_PROXY_CLIENT_IP,
+            Const.HTTP_CLIENT_IP,
+            Const.HTTP_X_FORWARDED_FOR,
+            Const.X_REAL_IP,
+            Const.X_REALIP,
+            Const.REMOTE_ADDR,
+            Const.UNKNOWN,
+            Const.LOCALHOST,
+            Const.ALL_IP
+        };
+    }
 }
