--- src/main/java/io/wisoft/wasabi/global/config/common/annotation/Anyone.java
@@ -1,13 +1,0 @@
-package io.wisoft.wasabi.global.config.common.annotation;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import static java.lang.annotation.ElementType.PARAMETER;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-@Retention(RUNTIME)
-@Target(PARAMETER)
-
-public @interface Anyone {
-}
