+++ src/main/java/io/wisoft/wasabi/global/config/web/filter/LogFilter.java
@@ -39,19 +39,7 @@ public class LogFilter extends OncePerRequestFilter {
      */
     private String getClientIp(final HttpServletRequest request) throws UnknownHostException {
 
-        final String[] headers = {
-                Const.X_FORWARDED_FOR,
-                Const.PROXY_CLIENT_IP,
-                Const.WL_PROXY_CLIENT_IP,
-                Const.HTTP_CLIENT_IP,
-                Const.HTTP_X_FORWARDED_FOR,
-                Const.X_REAL_IP,
-                Const.X_REALIP,
-                Const.REMOTE_ADDR,
-                Const.UNKNOWN,
-                Const.LOCALHOST,
-                Const.ALL_IP
-        };
+        final String[] headers = Const.getClientIpHeaders();
 
         String ip = Strings.EMPTY;
