+++ src/main/java/io/wisoft/wasabi/global/config/web/resolver/Anyone.java
@@ -1,7 +1,4 @@
-package io.wisoft.wasabi.global.config.common.annotation;
-
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotNull;
+package io.wisoft.wasabi.global.config.web.resolver;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
@@ -9,9 +6,8 @@ import java.lang.annotation.Target;
 import static java.lang.annotation.ElementType.PARAMETER;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
-@MemberId
-@Valid @NotNull
 @Retention(RUNTIME)
 @Target(PARAMETER)
-public @interface LoginRequired {
+
+public @interface Anyone {
 }
