+++ src/main/java/io/wisoft/wasabi/global/config/web/validator/FileExtensionValidator.java
@@ -1,4 +1,4 @@
-package io.wisoft.wasabi.global.config.common.annotation;
+package io.wisoft.wasabi.global.config.web.validator;
 
 import io.wisoft.wasabi.global.config.common.Const;
 import jakarta.validation.ConstraintValidator;
