+++ src/test/java/io/wisoft/wasabi/domain/auth/AuthControllerTest.java
@@ -11,8 +11,8 @@ import io.wisoft.wasabi.domain.auth.dto.SignupRequest;
 import io.wisoft.wasabi.domain.auth.dto.SignupResponse;
 import io.wisoft.wasabi.domain.member.Member;
 import io.wisoft.wasabi.domain.member.Part;
-import io.wisoft.wasabi.global.config.common.annotation.AnyoneResolver;
-import io.wisoft.wasabi.global.config.common.annotation.MemberIdResolver;
+import io.wisoft.wasabi.global.config.web.resolver.AnyoneResolver;
+import io.wisoft.wasabi.global.config.web.resolver.MemberIdResolver;
 import io.wisoft.wasabi.global.config.common.jwt.JwtTokenProvider;
 import io.wisoft.wasabi.global.config.web.interceptor.AdminInterceptor;
 import io.wisoft.wasabi.global.config.web.response.ResponseAspect;
