+++ src/test/java/io/wisoft/wasabi/domain/board/BoardControllerTest.java
@@ -13,8 +13,8 @@ import io.wisoft.wasabi.domain.member.Member;
 import io.wisoft.wasabi.domain.member.Role;
 import io.wisoft.wasabi.domain.tag.Tag;
 import io.wisoft.wasabi.global.config.common.Const;
-import io.wisoft.wasabi.global.config.common.annotation.AnyoneResolver;
-import io.wisoft.wasabi.global.config.common.annotation.MemberIdResolver;
+import io.wisoft.wasabi.global.config.web.resolver.AnyoneResolver;
+import io.wisoft.wasabi.global.config.web.resolver.MemberIdResolver;
 import io.wisoft.wasabi.global.config.common.jwt.JwtTokenProvider;
 import io.wisoft.wasabi.global.config.web.interceptor.AdminInterceptor;
 import io.wisoft.wasabi.global.config.web.response.ResponseAspect;
