+++ src/test/java/io/wisoft/wasabi/domain/like/LikeRepositoryTest.java
@@ -79,48 +79,6 @@ class LikeRepositoryTest {
     }
 
     @Nested
-    @DisplayName("좋아요 취소")
-    class CancelLike {
-
-        @DisplayName("요청 시 정상적으로 삭제되어야 한다.")
-        @ParameterizedTest
-        @AutoSource
-        @Customization(BoardCompositeCustomizer.class)
-        void cancel_like(final Member member,
-                         final Board board) {
-
-            // given
-            final Like like = init(member, board);
-
-            likeRepository.save(like);
-
-            // when
-            final int result = likeRepository.deleteByMemberIdAndBoardId(
-                    member.getId(),
-                    board.getId()
-            );
-
-            // then
-            assertThat(result).isOne();
-        }
-
-        @DisplayName("존재하지 않는 데이터 삭제시 아무것도 삭제되지 않는다.")
-        @ParameterizedTest
-        @AutoSource
-        void cancel_like_fail(final Long memberId,
-                              final Long boardId) {
-
-            // given
-
-            // when
-            final int result = likeRepository.deleteByMemberIdAndBoardId(memberId, boardId);
-
-            // then
-            assertThat(result).isZero();
-        }
-    }
-
-    @Nested
     @DisplayName("좋아요 조회")
     class FindLike {
 
@@ -162,43 +120,6 @@ class LikeRepositoryTest {
 
             // then
             assertThat(result).isEmpty();
-        }
-    }
-
-    @Nested
-    @DisplayName("좋아요 상태 조회")
-    class GetLikeStatus {
-
-        @DisplayName("BoardId 조회 시 정상적으로 좋아요수가 조회되어야 한다.")
-        @ParameterizedTest
-        @AutoSource
-        @Customization(BoardCompositeCustomizer.class)
-        void count_like_by_board_id(final Member member,
-                                    final Board board) {
-
-            //given
-            final Like like = init(member, board);
-            likeRepository.save(like);
-
-            //when
-            int result = likeRepository.countByBoardId(board.getId());
-
-            //then
-            assertThat(result).isOne();
-        }
-
-        @DisplayName("BoardId 조회 시 데이터가 없다면 빈값이 조회되어야 한다.")
-        @ParameterizedTest
-        @AutoSource
-        void count_like_by_board_id_fail(final Long boardId) {
-
-            //given
-
-            //when
-            int result = likeRepository.countByBoardId(boardId);
-
-            //then
-            assertThat(result).isZero();
         }
     }
 }
